refactor: Address PR review feedback
- Extract MetricsInterceptor to shared/pkg/interceptors/ (eliminates duplication)
- Add testing.Short() guard to GormDBChecker_Name test
- Clarify ADR about position-keeping repository/ migration status
- Document shared interceptors (MetricsInterceptor, RecoveryUnaryInterceptor)
