+++ services/financial-accounting/observability/health.go
@@ -1,3 +1,4 @@
+// Package observability provides health checks for the financial-accounting service.
 package observability
 
 import (
