+++ shared/pkg/interceptors/metrics.go
@@ -1,5 +1,4 @@
-// Package observability provides Prometheus metrics and health checks for the financial-accounting service.
-package observability
+package interceptors
 
 import (
 	"context"
@@ -11,6 +10,7 @@ import (
 )
 
 // MetricsInterceptor creates a gRPC unary interceptor that records Prometheus metrics.
+// It tracks request counts (by method and status code) and request duration.
 func MetricsInterceptor(requestsTotal *prometheus.CounterVec, requestDuration *prometheus.HistogramVec) grpc.UnaryServerInterceptor {
 	return func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
 		start := time.Now()
